requirements.txt
django
pymysql
yfinance
django-graphos